requirements.txt
backports.statistics==0.1.0
beautifulsoup4==4.10.0
Cython==0.29.27
dl==0.1.0
ipaddr==2.2.0
keyring==23.0.1
matplotlib==3.5.1
mypy_extensions==0.4.3
nltk==3.6.2
numpy
ordereddict==1.1
pandas==1.2.3
plotly==5.5.0
pydub==0.24.0
python-dotenv==0.19.2
pytube==11.0.2
schedule==1.1.0
scikit_learn==1.0.2
seaborn==0.11.2
selenium==3.141.0
SpeechRecognition==3.8.1
typing_extensions==4.0.1
yellowbrick==1.3.post1
